refactor: 将SDK核心方法从__init__移动到SDK类中
